+++ src/main/java/org/braekpo1nt/mctmanager/ui/UIException.java
@@ -1,4 +1,4 @@
-package org.braekpo1nt.mctmanager.ui.topbar;
+package org.braekpo1nt.mctmanager.ui;
 
 public class UIException extends RuntimeException {
     public UIException(String message) {
