+++ src/main/java/org/braekpo1nt/mctmanager/ui/UIUtils.java
@@ -8,7 +8,6 @@ import net.kyori.adventure.title.Title;
 import org.braekpo1nt.mctmanager.Main;
 import org.braekpo1nt.mctmanager.io.IOUtils;
 import org.braekpo1nt.mctmanager.ui.maps.ImageMapRenderer;
-import org.braekpo1nt.mctmanager.ui.topbar.UIException;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.World;
